To handle bad Cineplex recordings
During cases where Cineplex recording skips, the timestamps are
conserved so we observe frames where the sampling rate deviates
drastically from 0.03. This is problematic because it misaligns tracking
and imaging movies.

The above fix locates these frames and replaces the Xpix, Ypix, and time
vectors with the values from the last good frame. It is then recommended
to run this function through the whole DVT file because everything will
be shifted by the number of frames that were bad.